refactor: decompose index.ts and extract secure image fetch
src/index.ts was a 1,464-line file mixing four unrelated concerns. Split
it into focused modules, leaving index.ts (257 lines) as a thin
orchestrator:

- errors.ts: MarkdownConversionError
- validation.ts: all input/option validation
- sectionBuilder.ts: style/section normalization + merging, enum
  resolvers, and header/footer/section-property builders
- tocBuilder.ts: TOC content + placeholder replacement
- documentStyles.ts: Title/Heading/Strong paragraph styles

Also split the 766-line imageRenderer.ts: the SSRF-hardened network code
(DNS pinning, private-IP blocking, redirect/time budgets) moves to
utils/secureImageFetch.ts, leaving imageRenderer focused on decoding and
embedding.

While moving code, collapsed three duplicated blocks into data-driven
loops with no output change: the six near-identical Heading1-6 styles,
the default/first/even header & footer assembly, and the per-level TOC
entry styling switch. Verified the generated DOCX XML (document, styles,
numbering, headers, footers) is byte-identical to before via snapshot diff.
